Factor requests out into concern
